interactive_video_booth/requirements.txt
opencv-python
numpy
websockets
Flask
PyYAML
Pillow
pytest
